backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
pydantic
deep-translator
beautifulsoup4
lxml
email-validator
aiohttp